Added option to compile primes using only MHS, added some unit tests
PrimeCompilator now can be initialized with the argument `use_mhs_only`, which defaults to False. Having this enabled skips the quine-mccluskey part of the algorithm and only uses mhs to find all primes.

Added some basic unit tests that are (for now) only used to test the new method of generating primes (as explained above)